+++ src/TestSuite/TestCase.php
@@ -15,8 +15,7 @@ declare(strict_types=1);
 namespace Thumber\Cake\TestSuite;
 
 use MeTools\TestSuite\TestCase as BaseTestCase;
-use Thumber\Cake\Utility\ThumbCreator;
-use Thumber\TestSuite\TestTrait;
+use Tools\Filesystem;
 
 /**
  * TestCase class
@@ -26,45 +25,16 @@ abstract class TestCase extends BaseTestCase
     use TestTrait;
 
     /**
-     * Asserts for a valid thumbnail url
-     * @param string $url Thumbnail url
-     * @param string $message The failure message that will be appended to the
-     *  generated message
+     * Called after every test method
      * @return void
-     * @since 1.4.0
+     * @throws \Symfony\Component\Filesystem\Exception\IOException
+     * @throws \Symfony\Component\Finder\Exception\DirectoryNotFoundException
+     * @throws \Tools\Exception\MethodNotExistsException
      */
-    public function assertThumbUrl(string $url, string $message = ''): void
+    protected function tearDown(): void
     {
-        self::assertMatchesRegularExpression('/^(http:\/\/localhost)?\/thumb\/[\w\d]+/', $url, $message);
-    }
+        Filesystem::instance()->unlinkRecursive(THUMBER_TARGET);
 
-    /**
-     * Returns an instance of `ThumbCreator`
-     * @param string $path Path of the image from which to create the
-     *  thumbnail. It can be a relative path (to APP/webroot/img), a full path
-     *  or a remote url
-     * @return \Thumber\Cake\Utility\ThumbCreator
-     */
-    protected function getThumbCreatorInstance(string $path = ''): ThumbCreator
-    {
-        return new ThumbCreator($path ?: '400x400.jpg');
-    }
-
-    /**
-     * Returns an instance of `ThumbCreator`, after calling `resize()` and
-     *  `save()` methods.
-     *
-     * It can be called passing only the array of options as first argument.
-     * @param string $path Path of the image from which to create the thumbnail.
-     *  It can be a full path or a remote url
-     * @param array $options Options for saving
-     * @return \Thumber\Cake\Utility\ThumbCreator
-     */
-    protected function getThumbCreatorInstanceWithSave(string $path = '', array $options = []): ThumbCreator
-    {
-        $thumbCreator = $this->getThumbCreatorInstance($path);
-        $thumbCreator->resize(200)->save($options);
-
-        return $thumbCreator;
+        parent::tearDown();
     }
 }