+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * the `ThumbCreator` class has been revised;
 * `ThumberCreator::getImageInstance()` and `ThumberCreator::save()` methods have been, because they would now coincide
   with the methods of the extended class;
+* all methods of `TestCase` class have been moved to the new `TestTrait`;
 * added tests for `ThumbManager` class.
 
 ## 1.10 branch